Switch from relative to absolute imports and fix circular dependency. Fixes #158.
